Association for Molecular Pathology
Next-Generation Sequencing: The AMP/ASCP certifies that the named individual has participated in this enduring activity on the date indicated above and is awarded the
indicated number of ASCP CMLE Credit(s). This activity has been planned and implemented by the Essential Areas and policies of the Accreditation Council for
Continuing Medical Education through the joint providership of the American Society for Clinical Pathology (ASCP) and the Association
for Molecular Pathology (AMP).