AI & Big Data: Data Science and Data Governance
Dr. Nambili Samuel has successfully completed the AI & Big Data: Data Science and Governance program and receiving a Certificate of Application from Mahidol University (MU), the oldest research institution in the world, established in 1888 and home to the renowned University and Siriraj Hospital in Bangkok, Thailand. This merit comes with transferable credits to any academic institution.